include: give the ported funcs.h what it needs, keeping our compat shims
sfpi_funcs.h now comes from main, and it references two things this
branch's headers did not have:

    sfpi_funcs.h:144: error: 'sfpi::nearby(vBool)' should have been
                      declared inside 'sfpi'
    sfpi_funcs.h:268: error: 'SM8' is not a member of 'sfpi::DataLayout'

Both are additive, so they are added rather than the files replaced.
That matters: most of what this branch's sfpi.h carries over main's is
main DELETING things we still implement -- sFloat16b(double), the
s2vFloat16a/b wrappers, vBool's conversion to vInt, and the
mixed-signedness comparison operators are all '= delete' upstream and
live here.  They are backward compatibility the corpus may rely on, and
taking main's sfpi.h wholesale would have removed them silently.

  sfpi_classes.h  DataLayout gains U8 and SM8, the 8-bit formats main's
                  load/store switches enumerate
  sfpi.h          nearby declared beside the other vBool operators
  lltt.h          main's instruction-buffer type (unsigned long, which
                  the rebased compiler requires -- ours had narrowed it
                  to uint32_t, a different type) with our typed setrwc
                  template kept

+++ include/lltt.h
@@ -15,10 +15,10 @@
 #define __builtin_rvtt_ttreplay(START, LENGTH, EXEC, RECORD) \
   __builtin_rvtt_ttreplay(lltt::__instrn_buffer, LENGTH, 0, 0, START, EXEC, RECORD)
 
-extern volatile uint32_t __instrn_buffer[];
+extern volatile unsigned long __instrn_buffer[];
 
 namespace lltt {
-constexpr inline volatile uint32_t *[[gnu::rvtt_reg_ptr]] __instrn_buffer = ::__instrn_buffer;
+constexpr inline volatile unsigned long *[[gnu::rvtt_reg_ptr]] __instrn_buffer = ::__instrn_buffer;
 
 enum ExecBool : bool {NoExec, Exec};
